Added Helper, Created static TotalSum class
